+++ express.js
@@ -1,4 +1,5 @@
 var express = require("express")
+  , crypto = require('crypto')
   , cookieSession = require('cookie-session')
   , passport = require('passport')
   , path = require('path')
@@ -33,7 +34,7 @@ app.use(function(req, res, next) {
   return next();
 });
 
-const secret = require('crypto').randomBytes(64).toString('hex');
+const secret = crypto.randomBytes(64).toString('hex');
 app.use(cookieSession({
   secret: secret,
   name: 'mage-session',
@@ -56,7 +57,6 @@ app.use(require('body-parser')({
   limit: '16mb',
   keepExtensions: true
 }));
-app.use(require('multer')());
 app.use(passport.initialize());
 app.use(passport.session());
 app.use(express.static(path.join(__dirname, 'public/dist')));