+++ routes/imports.js
@@ -2,22 +2,25 @@ module.exports = function(app, security) {
   var api = require('../api')
     , access = require('../access')
     , fs = require('fs-extra')
+    , {default: upload} = require('../upload')
     , toGeoJson = require('../utilities/togeojson');
 
   var passport = security.authentication.passport;
 
   function verifyLayer(req, res, next) {
     if (req.layer.type !== 'Feature') {
-      return res.status(400).send("Cannot import data, layer type is not 'Static'");
+      return res.status(400).send('Cannot import data, layer type is not "Static"');
+    }
+
+    if (req.file.mimetype !== 'application/vnd.google-earth.kml+xml') {
+      return res.status(400).send('Cannot import data, upload file must be KML');
     }
 
     return next();
   }
 
   function readImportFile(req, res, next) {
-    // TODO at some point open file and determine type (KML, shapefile, geojson, csv)
-
-    fs.readFile(req.files.file.path, 'utf8', function(err, data) {
+    fs.readFile(req.file.path, 'utf8', function(err, data) {
       req.importData = data;
       return next(err);
     });
@@ -27,6 +30,7 @@ module.exports = function(app, security) {
     '/api/layers/:layerId/kml',
     passport.authenticate('bearer'),
     access.authorize('CREATE_LAYER'),
+    upload.single('file'),
     verifyLayer,
     readImportFile,
     function(req, res, next) {
@@ -35,8 +39,8 @@ module.exports = function(app, security) {
         .then(newFeatures => {
           var response = {
             files: [{
-              name: req.files.file.originalname,
-              size: req.files.file.size,
+              name: req.file.originalname,
+              size: req.file.size,
               features: newFeatures.length
             }]
           };