+++ test/user/userUpdateTest.js
@@ -19,19 +19,14 @@ require('sinon-mongoose');
 
 describe("user update tests", function() {
 
-  var sandbox;
-  before(function() {
-    sandbox = sinon.sandbox.create();
-  });
-
   afterEach(function() {
-    sandbox.restore();
+    sinon.restore();
   });
 
   var userId = mongoose.Types.ObjectId();
   function mockTokenWithPermission(permission) {
     var permissions = Array.isArray(permission) ? permission : [permission];
-    sandbox.mock(TokenModel)
+    sinon.mock(TokenModel)
       .expects('findOne')
       .withArgs({token: "12345"})
       .chain('populate', 'userId')
@@ -61,18 +56,18 @@ describe("user update tests", function() {
       }
     };
 
-    sandbox.mock(TokenModel)
+    sinon.mock(TokenModel)
       .expects('findOne')
       .withArgs({token: "12345"})
       .chain('populate', 'userId')
       .chain('exec')
       .yields(null, token);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('save')
       .yields(null, mockUser);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('populate')
       .yields(null, mockUser);
 
@@ -108,18 +103,18 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findOne')
       .withArgs({ username: 'test' })
       .chain('populate', 'roleId')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(UserModel.prototype)
+    sinon.mock(UserModel.prototype)
       .expects('validPassword')
       .yields(null, true);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('save')
       .resolves(mockUser);
 
@@ -155,18 +150,18 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findOne')
       .withArgs({ username: 'test' })
       .chain('populate', 'roleId')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(UserModel.prototype)
+    sinon.mock(UserModel.prototype)
       .expects('validPassword')
       .yields(null, true);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('save')
       .resolves(mockUser);
 
@@ -204,16 +199,16 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('save')
       .yields(null, mockUser);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('populate')
       .yields(null, mockUser);
 
@@ -240,7 +235,7 @@ describe("user update tests", function() {
   });
 
   it('should update user role with UPDATE_USER_ROLE', function(done) {
-    sandbox.mock(TokenModel)
+    sinon.mock(TokenModel)
       .expects('findOne')
       .withArgs({token: "12345"})
       .chain('populate', 'userId')
@@ -259,12 +254,12 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(User)
+    sinon.mock(User)
       .expects('updateUser')
       .withArgs(sinon.match.has('roleId', roleId))
       .yields(null, mockUser);
@@ -305,12 +300,12 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(User)
+    sinon.mock(User)
       .expects('updateUser')
       .withArgs(sinon.match.has('authentication', sinon.match.has('password', 'passwordpassword')))
       .yields(null, mockUser);
@@ -352,12 +347,12 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(User)
+    sinon.mock(User)
       .expects('updateUser')
       .withArgs(sinon.match.has('authentication', sinon.match.has('password', sinon.match.typeOf("undefined"))))
       .yields(null, mockUser);
@@ -400,12 +395,12 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(User)
+    sinon.mock(User)
       .expects('updateUser')
       .withArgs(sinon.match.has('roleId', undefined))
       .yields(null, mockUser);
@@ -447,12 +442,12 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(User)
+    sinon.mock(User)
       .expects('updateUser')
       .withArgs(sinon.match({active: true}))
       .yields(null, mockUser);
@@ -482,12 +477,12 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(User)
+    sinon.mock(User)
       .expects('updateUser')
       .withArgs(sinon.match({enabled: false}))
       .yields(null, mockUser);
@@ -527,20 +522,20 @@ describe("user update tests", function() {
       callback(null, mockUser);
     };
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(UserModel.collection)
+    sinon.mock(UserModel.collection)
       .expects('update')
       .yields(null, {});
 
-    sandbox.mock(UserModel.collection)
+    sinon.mock(UserModel.collection)
       .expects('findOne')
       .yields(null, null);
 
-    const tokenStub = sandbox.mock(TokenModel)
+    const tokenStub = sinon.mock(TokenModel)
       .expects('remove')
       .withArgs(sinon.match({userId: id}))
       .yields(null);
@@ -583,20 +578,20 @@ describe("user update tests", function() {
       callback(null, mockUser);
     };
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
 
-    sandbox.mock(UserModel.collection)
+    sinon.mock(UserModel.collection)
       .expects('update')
       .yields(null, {});
 
-    sandbox.mock(UserModel.collection)
+    sinon.mock(UserModel.collection)
       .expects('findOne')
       .yields(null, null);
 
-    const tokenStub = sandbox.mock(TokenModel)
+    const tokenStub = sinon.mock(TokenModel)
       .expects('remove')
       .withArgs(sinon.match({userId: id}))
       .yields(null);
@@ -630,7 +625,7 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
@@ -667,7 +662,7 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
@@ -709,18 +704,18 @@ describe("user update tests", function() {
       }
     };
 
-    sandbox.mock(TokenModel)
+    sinon.mock(TokenModel)
       .expects('findOne')
       .withArgs({token: "12345"})
       .chain('populate', 'userId')
       .chain('exec')
       .yields(null, token);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('save')
       .yields(null, mockUser);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('populate')
       .yields(null, mockUser);
 
@@ -762,7 +757,7 @@ describe("user update tests", function() {
       }
     };
 
-    sandbox.mock(TokenModel)
+    sinon.mock(TokenModel)
       .expects('findOne')
       .withArgs({token: "12345"})
       .chain('populate', 'userId')
@@ -801,18 +796,18 @@ describe("user update tests", function() {
       }
     };
 
-    sandbox.mock(TokenModel)
+    sinon.mock(TokenModel)
       .expects('findOne')
       .withArgs({token: "12345"})
       .chain('populate', 'userId')
       .chain('exec')
       .yields(null, token);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('save')
       .yields(null, mockUser);
 
-    sandbox.mock(mockUser)
+    sinon.mock(mockUser)
       .expects('populate')
       .yields(null, mockUser);
 
@@ -848,7 +843,7 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
@@ -858,7 +853,7 @@ describe("user update tests", function() {
       name: 'Mock Event'
     });
 
-    sandbox.mock(EventModel)
+    sinon.mock(EventModel)
       .expects('findById')
       .twice()
       .onFirstCall()
@@ -866,7 +861,7 @@ describe("user update tests", function() {
       .onSecondCall()
       .yields(null, mockEvent);
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findByIdAndUpdate')
       .withArgs(userId, {recentEventIds: [1]}, {new: true})
       .yields(null, mockUser);
@@ -908,14 +903,14 @@ describe("user update tests", function() {
       }
     };
 
-    sandbox.mock(TokenModel)
+    sinon.mock(TokenModel)
       .expects('findOne')
       .withArgs({token: "12345"})
       .chain('populate', 'userId')
       .chain('exec')
       .yields(null, token);
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
@@ -928,7 +923,7 @@ describe("user update tests", function() {
       acl: eventAcl
     });
 
-    sandbox.mock(EventModel)
+    sinon.mock(EventModel)
       .expects('findById')
       .twice()
       .onFirstCall()
@@ -936,7 +931,7 @@ describe("user update tests", function() {
       .onSecondCall()
       .yields(null, mockEvent);
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findByIdAndUpdate')
       .withArgs(userId, {recentEventIds: [6,5,4,3,2]}, {new: true})
       .yields(null, mockUser);
@@ -968,7 +963,7 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
@@ -981,7 +976,7 @@ describe("user update tests", function() {
       acl: eventAcl
     });
 
-    sandbox.mock(EventModel)
+    sinon.mock(EventModel)
       .expects('findById')
       .twice()
       .onFirstCall()
@@ -989,7 +984,64 @@ describe("user update tests", function() {
       .onSecondCall()
       .yields(null, mockEvent);
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
+      .expects('findByIdAndUpdate')
+      .withArgs(userId, {recentEventIds: [1]}, {new: true})
+      .yields(null, mockUser);
+
+    request(app)
+      .post('/api/users/' + userId.toString() + '/events/1/recent' )
+      .set('Accept', 'application/json')
+      .set('Authorization', 'Bearer 12345')
+      .expect(200)
+      .expect('Content-Type', /json/)
+      .expect(function(res) {
+        var user = res.body;
+        should.exist(user);
+        user.should.have.property('id').that.equals(userId.toString());
+      })
+      .end(done);
+  });
+
+  it('should add recent event for user in event', function(done) {
+    mockTokenWithPermission('NO_ADMIN_PERMISSION');
+
+    var mockUser = new UserModel({
+      _id: userId,
+      username: 'test',
+      displayName: 'test',
+      active: true,
+      authentication: {
+        type: 'local'
+      }
+    });
+
+    sinon.mock(UserModel)
+      .expects('findById')
+      .chain('exec')
+      .yields(null, mockUser);
+
+    var mockEvent1 = new EventModel({
+      _id: 1,
+      name: 'Mock Event 12345',
+      acl: {}
+    });
+
+    sinon.mock(EventModel)
+      .expects('findById')
+      .twice()
+      .onFirstCall()
+      .yields(null, mockEvent1);
+
+    sinon.mock(EventModel)
+      .expects('populate')
+      .yields(null, {
+        teamIds: [{
+          userIds: [userId]
+        }]
+      });
+
+    sinon.mock(UserModel)
       .expects('findByIdAndUpdate')
       .withArgs(userId, {recentEventIds: [1]}, {new: true})
       .yields(null, mockUser);
@@ -1021,7 +1073,7 @@ describe("user update tests", function() {
       }
     });
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findById')
       .chain('exec')
       .yields(null, mockUser);
@@ -1032,7 +1084,7 @@ describe("user update tests", function() {
       acl: {}
     });
 
-    sandbox.mock(EventModel)
+    sinon.mock(EventModel)
       .expects('findById')
       .twice()
       .onFirstCall()
@@ -1040,7 +1092,7 @@ describe("user update tests", function() {
       .onSecondCall()
       .yields(null, mockEvent);
 
-    sandbox.mock(UserModel)
+    sinon.mock(UserModel)
       .expects('findByIdAndUpdate')
       .withArgs(userId, {recentEventIds: [1]}, {new: true})
       .yields(null, mockUser);